pyproject.toml
[tool.poetry]
name = "01-Setup-DBT"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.2"
psycopg2-binary = "^2.9.3"
google-api-python-client = "^2.52.0"
dbt-bigquery = "^1.1.0"
pyarrow = "^8.0.0"
google-cloud-bigquery = "^2.0.0"


[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
ipython = "^8.4.0"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 160
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 160
multi_line_output = 3
include_trailing_comma = true

[virtualenvs]
in-project = true
